fix(trading): use BigDecimal throughout FX rate resolution [#42]
Address Gemini code review on PR #49: Float literals and .to_f in
financial calculations can accumulate precision errors.

Changes:
- ExecuteTradeContract: optional(:fx_rate_at_execution).maybe(:decimal)
  instead of :float — preserves BigDecimal precision from input.
- ExecuteTrade#resolve_fx_rate: replace 1.0/0.0 literals with BigDecimal(1);
  drop .to_f conversions on values from FxRate.convert (already BigDecimal).
  Inverse rate now BigDecimal(1) / inverse for precise division.

Tests: 2201 examples, 0 failures (no spec changes needed — DB column is
already decimal and round-trip comparisons handle Float/BigDecimal coercion).

+++ app/contexts/trading/contracts/execute_trade_contract.rb
@@ -9,7 +9,7 @@ module Trading
         optional(:fee).maybe(:float)
         optional(:executed_at).maybe(:string)
         optional(:currency).maybe(:string, included_in?: Asset::SUPPORTED_CURRENCIES)
-        optional(:fx_rate_at_execution).maybe(:float)
+        optional(:fx_rate_at_execution).maybe(:decimal)
       end
 
       rule(:shares) do
